added logger logic and fixed bugs 1.4

+++ src/main/java/com/trans/service/impl/CargoServiceImpl.java
@@ -2,9 +2,12 @@ package com.trans.service.impl;
 
 import com.trans.dto.UserDTO;
 import com.trans.model.Cargo;
+import com.trans.model.Order;
 import com.trans.model.User;
+import com.trans.model.enums.OrderStatus;
 import com.trans.repository.CargoRepository;
 import com.trans.service.CargoService;
+import com.trans.service.OrderService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.*;
@@ -21,11 +24,13 @@ import java.util.stream.Collectors;
 public class CargoServiceImpl implements CargoService {
 
     private final CargoRepository cargoRepository;
+
     private final ModelMapper modelMapper;
 
     @Autowired
     public CargoServiceImpl(CargoRepository cargoRepository, ModelMapper modelMapper) {
         this.cargoRepository = cargoRepository;
+
         this.modelMapper = modelMapper;
     }
 
@@ -50,10 +55,17 @@ public class CargoServiceImpl implements CargoService {
     @Override
     public boolean deleteById(int id) {
         Cargo byId = findById(id);
-        byId.getOrderList().forEach(order -> order.setCargo(null));
-        byId.setOrderList(null);
-        save(byId);
+        boolean notReadyToDelete = byId.getOrderList().stream()
+                .anyMatch(order -> order.getStatus() == OrderStatus.ACTIVE || order.getStatus() == OrderStatus.WAITING);
+        if(notReadyToDelete){
+            return false;
+        }
+
+        if(byId.getOrderList().isEmpty()){
         cargoRepository.deleteById(id);
+        }else {
+            byId.setDelete(true);
+        }
         return true;
     }
